requirements.txt
pyotp==2.6.0
pycryptodome==3.11.0
Pillow==9.0.1
qrcode==7.3
deep_translator==1.9.1